requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
email_validator==1.3.1
cryptography==41.0.2
python-dotenv==1.0.0
Werkzeug==2.2.3